Added tests for DataStorange
